Untangle frontend from ElementInventoryService
